requirements.txt
# dbt packages
dbt-core==1.9.1
dbt-postgres==1.9.0

# Common dependency for dbt-postgres and potentially your scripts
psycopg2-binary~=2.9.0

# Dependencies for your Python scripts in the 'scripts/' directory 
# (e.g., SourceData.py, jsonParser.py, load_to_postgres.py).
# Ensure these are needed if you run these scripts directly in the 'app' container 
# (i.e., not via Airflow, which uses requirements-airflow.txt).
ijson~=3.3.0          # For jsonParser.py
pandas~=2.2.0         # For general data manipulation in scripts
pyarrow~=20.0.0       # For Parquet handling or as a pandas accelerator
requests~=2.32.0      # If scripts like SourceData.py make direct HTTP calls

# Web interface dependencies
flask~=3.0.0          # Web framework for the web interface
gunicorn~=21.2.0      # WSGI server for production deployment

# Added for Flask reloader
watchdog